+++ pkg/storage/ingest/config.go
@@ -277,7 +277,9 @@ func (cfg *KafkaConfig) GetConsumerGroup(instanceID string, partitionID int32) s
 // MigrationConfig holds the configuration used to migrate Mimir to ingest storage. This config shouldn't be
 // set for any other reason.
 type MigrationConfig struct {
-	DistributorSendToIngestersEnabled bool `yaml:"distributor_send_to_ingesters_enabled"`
+	DistributorSendToIngestersEnabled bool          `yaml:"distributor_send_to_ingesters_enabled"`
+	IgnoreIngestStorageErrors         bool          `yaml:"ignore_ingest_storage_errors"`
+	IngestStorageMaxWaitTime          time.Duration `yaml:"ingest_storage_max_wait_time"`
 }
 
 func (cfg *MigrationConfig) RegisterFlags(f *flag.FlagSet) {
@@ -286,4 +288,6 @@ func (cfg *MigrationConfig) RegisterFlags(f *flag.FlagSet) {
 
 func (cfg *MigrationConfig) RegisterFlagsWithPrefix(prefix string, f *flag.FlagSet) {
 	f.BoolVar(&cfg.DistributorSendToIngestersEnabled, prefix+"distributor-send-to-ingesters-enabled", false, "When both this option and ingest storage are enabled, distributors write to both Kafka and ingesters. A write request is considered successful only when written to both backends.")
+	f.BoolVar(&cfg.IgnoreIngestStorageErrors, prefix+"ignore-ingest-storage-errors", false, "When enabled, errors writing to ingest storage are logged but do not affect write success or quorum. When disabled, write requests fail if ingest storage write fails.")
+	f.DurationVar(&cfg.IngestStorageMaxWaitTime, prefix+"ingest-storage-max-wait-time", 0, "The maximum time a write request that goes through the ingest storage waits before it times out. Set to `0` to disable the timeout.")
 }